--- a/lib/taric_importer.rb
+++ b/lib/taric_importer.rb
@@ -2,7 +2,6 @@ require 'nokogiri'
 
 require 'taric_importer/transaction'
 require 'taric_importer/record_processor'
-require 'taric_importer/helpers/string_helper'
 
 class TaricImporter < TariffImporter
   class ImportException < StandardError