--- a/lib/taric_importer.rb
+++ b/lib/taric_importer.rb
@@ -2,6 +2,7 @@ require 'nokogiri'
 
 require 'taric_importer/transaction'
 require 'taric_importer/record_processor'
+require 'taric_importer/xml_parser'
 
 class TaricImporter < TariffImporter
   class ImportException < StandardError
@@ -16,34 +17,15 @@ class TaricImporter < TariffImporter
   class UnknownOperationError < ImportException
   end
 
-  cattr_accessor :opening_element
-  self.opening_element = 1
-
-  cattr_accessor :transaction_node
-  self.transaction_node = 'env:transaction'
-
   def import
-    xml = nil
+    XmlParser::Reader.new(path, "record", XmlProcessor.new).parse
+  end
 
-    ActiveSupport::Notifications.instrument("taric_imported.tariff_importer", path: path) do
-      begin
-        handler = File.open(path, "r")
-        reader = Nokogiri::XML::Reader(handler, nil, nil, Nokogiri::XML::ParseOptions::RECOVER | Nokogiri::XML::ParseOptions::NOERROR | Nokogiri::XML::ParseOptions::NONET)
-        reader.each do |node|
-          if node.name == self.transaction_node && node.node_type == self.opening_element
-            xml = Nokogiri::XML(node.outer_xml).remove_namespaces!
-            transaction = Transaction.new(Hash.from_xml(xml.to_s), issue_date)
-            transaction.persist
-            transaction.validate
-          end
-        end
-      rescue StandardError => exception
-        ActiveSupport::Notifications.instrument("taric_failed.tariff_importer",
-          exception: exception,
-          xml: xml
-        )
-        raise ImportException.new
-      end
+  class XmlProcessor
+    def process_xml_node property
+      transaction = Transaction.new(property, nil)
+      transaction.persist
+      transaction.validate
     end
   end
 end