--- a/lib/taric_importer/record_processor/create_operation.rb
+++ b/lib/taric_importer/record_processor/create_operation.rb
@@ -2,7 +2,7 @@ class TaricImporter < TariffImporter
   class RecordProcessor
     class CreateOperation < Operation
       def call
-        klass.create(attributes)
+        klass.new(attributes).save(validate: false, transaction: false)
       end
 
       def to_oplog_operation