--- a/spec/integration/chief_transformer/custom_spec.rb
+++ b/spec/integration/chief_transformer/custom_spec.rb
@@ -40,18 +40,18 @@ describe 'CHIEF: Custom scenarios' do
       # As goods_nomenclature now does not span the validity period of national_measure (ME8)
       def perform_transaction()
         transaction_record = {
-          'transaction.id' => transaction_id,
-          'record.code' => '400',
-          'subrecord.code' => '00',
-          'record.sequence.number' => '199',
-          'update.type' => '1',
-          'goods.nomenclature' => {
-            'goods.nomenclature.sid' => '70180',
-            'goods.nomenclature.item.id' => '1604141620',
-            'producline.suffix' => '80',
-            'validity.start.date' => '1998-07-14',
-            'validity.end.date' => '2011-12-31',
-            'statistical.indicator' => '0'
+          'transaction_id' => transaction_id,
+          'record_code' => '400',
+          'subrecord_code' => '00',
+          'record_sequence_number' => '199',
+          'update_type' => '1',
+          'goods_nomenclature' => {
+            'goods_nomenclature_sid' => '70180',
+            'goods_nomenclature_item_id' => '1604141620',
+            'producline_suffix' => '80',
+            'validity_start_date' => '1998-07-14',
+            'validity_end_date' => '2011-12-31',
+            'statistical_indicator' => '0'
           }
 
         }
@@ -113,17 +113,17 @@ describe 'CHIEF: Custom scenarios' do
       # Goods Nomenclature deletion Makes national measures invalid.
       def perform_transaction()
         transaction_record = {
-          'transaction.id' => transaction_id,
-          'record.code' => '400',
-          'subrecord.code' => '00',
-          'record.sequence.number' => '00',
-          'update.type' => '2',
-          'goods.nomenclature' => {
-            'goods.nomenclature.sid' => '96196',
-            'goods.nomenclature.item.id' => '0305720040',
-            'producline.suffix' => '80',
-            'validity.start.date' => '2012-01-01',
-            'statistical.indicator' => '0'
+          'transaction_id' => transaction_id,
+          'record_code' => '400',
+          'subrecord_code' => '00',
+          'record_sequence_number' => '00',
+          'update_type' => '2',
+          'goods_nomenclature' => {
+            'goods_nomenclature_sid' => '96196',
+            'goods_nomenclature_item_id' => '0305720040',
+            'producline_suffix' => '80',
+            'validity_start_date' => '2012-01-01',
+            'statistical_indicator' => '0'
           }
         }