--- a/spec/unit/taric_importer/record_processor_spec.rb
+++ b/spec/unit/taric_importer/record_processor_spec.rb
@@ -6,14 +6,14 @@ require 'taric_importer/record_processor/create_operation'
 
 describe TaricImporter::RecordProcessor do
   let(:record_hash) {
-    {"transaction.id"=>"31946",
-     "record.code"=>"130",
-     "subrecord.code"=>"05",
-     "record.sequence.number"=>"1",
-     "update.type"=>"3",
-     "language.description"=>
-      {"language.code.id"=>"FR",
-       "language.id"=>"EN",
+    {"transaction_id"=>"31946",
+     "record_code"=>"130",
+     "subrecord_code"=>"05",
+     "record_sequence_number"=>"1",
+     "update_type"=>"3",
+     "language_description"=>
+      {"language_code_id"=>"FR",
+       "language_id"=>"EN",
        "description"=>"French"}}
   }