--- a/spec/unit/taric_importer/transaction_spec.rb
+++ b/spec/unit/taric_importer/transaction_spec.rb
@@ -4,23 +4,17 @@ require 'taric_importer/transaction'
 
 describe TaricImporter::Transaction do
   let(:record) {
-    {"transaction"=>
-      {"id"=>"1",
-       "app.message"=>
-        {"id"=>"8",
-         "transmission"=>
-          {"record"=>
-            {"transaction.id"=>"31946",
-             "record.code"=>"130",
-             "subrecord.code"=>"05",
-             "record.sequence.number"=>"1",
-             "update.type"=>"3",
-             "language.description"=>
-              {"language.code.id"=>"FR",
-               "language.id"=>"EN",
-               "description"=>"French"}}}}}}
+    {"transaction_id"=>"31946",
+     "record_code"=>"130",
+     "subrecord_code"=>"05",
+     "record_sequence_number"=>"1",
+     "update_type"=>"3",
+     "language_description"=>
+      {"language_code_id"=>"FR",
+       "language_id"=>"EN",
+       "description"=>"French"}}
   }
-  let(:transaction_date) { Date.today }
+  let(:transaction_date) { Date.current }
 
   describe 'initialization' do
     context 'invalid record structure provided' do