… Gemfile.lock
… spec/unit/taric_importer_spec.rb