@matthewford 2016-03-01
Happy about this overall - we can assume that the system is valid on the date of the seed files and yes we can run a validation over all the data anyway.