@matthewford 2016-03-01
One thing to note: transactions, I think this was so that we don't leave the db in an inconsistant state - should a daily sync fail we should rollback all changes. 

However as the tariff is a point in time db, we can also achieve this by just deleting all data from that import date (effectively what we do when we do a rollback). Basically we want to make sure the Last updated date published is correct, and it the sync fails we dont display half synced data.