@theharq 2016-03-01
PR updated, no pending tests, many commits squashed to more a little more meaningful commit messages